chore(dependencies): update dependencies in pyproject.toml
Update Python version constraint to "\<4.0" to ensure compatibility with
future Python releases. Upgrade development dependencies: - mypy to
"\^1.17.1" - ruff to "\^0.12.8" - sphinx-autodoc-typehints to
"\^3.2.0" - behave to "\^1.3.0" - coverage to "\^7.10.2" - pytest to
"\^8.4.1" - pytest-cov to "\^6.2.1" - pytest-mock to "\^3.14.1" - tox to
"\^4.28.4"

These updates ensure the project remains up-to-date with the latest
features and bug fixes in these tools.